Controller Updates

+++ Tim26-ISS/src/main/java/com/example/demo/repository/RejectionMessageRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.RejectionMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RejectionMessageRepository extends JpaRepository<RejectionMessage, Long> {
+public interface RejectionMessageRepository extends JpaRepository<RejectionMessage, Integer> {
 }
