+++ Tim26-ISS/src/main/java/com/example/demo/repository/WarningRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.Warning;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WarningRepository extends JpaRepository<Warning, Long> {
+public interface WarningRepository extends JpaRepository<Warning, Integer> {
 }
